+++ src/main/php/web/Status.class.php
@@ -1,10 +1,12 @@
 <?php namespace web;
 
+/** @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Status */
 class Status {
   private static $messages= [
     100 => 'Continue',
     101 => 'Switching Protocols',
     102 => 'Processing',
+    103 => 'Early Hints',
     200 => 'OK',
     201 => 'Created',
     202 => 'Accepted',